download-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
minio==7.1.0
python-dotenv==0.19.0
